chore: extract test-util/port.ts
