+++ test-util/mod.ts
@@ -1,2 +1,3 @@
 export * from "./config.ts";
 export * from "./pairs.ts";
+export * from "./port.ts";
